+++ walkers/src/http_tiles.rs
@@ -5,7 +5,7 @@ use reqwest_middleware::ClientWithMiddleware;
 use crate::io::Fetch;
 use crate::io::http::http_client;
 use crate::io::tiles_io::TilesIo;
-use crate::projector::Projection;
+use crate::projector::{MercatorProjection, Projection};
 use crate::sources::{Attribution, TileSource};
 use crate::style::Style;
 use crate::tiles::{EguiTileFactory, interpolate_from_lower_zoom};
@@ -13,7 +13,7 @@ use crate::{HttpOptions, TilePiece, Tiles};
 use crate::{Stats, TileId};
 
 /// Downloads the tiles via HTTP. It must persist between frames.
-pub struct HttpTiles<P: Projection> {
+pub struct HttpTiles<P: Projection = MercatorProjection> {
     attribution: Attribution,
     tiles_io: TilesIo,
     projection: P,