+++ walkers/src/pmtiles.rs
@@ -1,7 +1,7 @@
 use crate::{
     TileId, TilePiece, Tiles,
     io::{Fetch, tiles_io::TilesIo},
-    projector::Projection,
+    projector::{MercatorProjection, Projection},
     sources::Attribution,
     style::Style,
     tiles::{EguiTileFactory, interpolate_from_lower_zoom},
@@ -22,7 +22,7 @@ const DEFAULT_TILE_SIZE: u32 = 1024;
 /// Provides tiles from a local PMTiles file.
 ///
 /// <https://docs.protomaps.com/guide/getting-started>
-pub struct PmTiles<P: Projection> {
+pub struct PmTiles<P: Projection = MercatorProjection> {
     tiles_io: TilesIo,
     tile_size: u32,
     projection: P,